subscription: try to make it a bit more efficient
Use a compile-time regexp rather than run-time, hopefully creating
less garbage to be collected.  Also make it be Unicode (although
we don't depend on this yet) and case-insensitive.  Using a
case-insensitive regexp allows us to defer downcasing the string
until we're certain we have a match, and only downcase the hashtag
that matched rather than the whole text of the post.